Move errors to `lib/grandpa/models` - `ErrBlockDoesNotExist` - `ErrVoterNotFound`
