+++ lib/grandpa/message_handler_test.go
@@ -14,7 +14,6 @@ import (
 	"github.com/ChainSafe/gossamer/lib/blocktree"
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/crypto/ed25519"
-	grandpaerrors "github.com/ChainSafe/gossamer/lib/grandpa/errors"
 	"github.com/ChainSafe/gossamer/lib/grandpa/models"
 	"github.com/ChainSafe/gossamer/lib/keystore"
 	"github.com/ChainSafe/gossamer/pkg/scale"
@@ -501,7 +500,7 @@ func TestVerifyJustification_InvalidAuthority(t *testing.T) {
 	}
 
 	err = h.verifyJustification(just, 77, gs.state.SetID, models.Precommit)
-	require.ErrorIs(t, err, grandpaerrors.ErrVoterNotFound)
+	require.ErrorIs(t, err, models.ErrVoterNotFound)
 }
 
 func TestMessageHandler_VerifyPreVoteJustification(t *testing.T) {
