+++ lib/grandpa/message_tracker_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/ChainSafe/gossamer/dot/types"
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/crypto/ed25519"
-	"github.com/ChainSafe/gossamer/lib/grandpa/errors"
 	"github.com/ChainSafe/gossamer/lib/grandpa/models"
 	"github.com/ChainSafe/gossamer/lib/keystore"
 
@@ -54,7 +53,7 @@ func TestMessageTracker_ValidateMessage(t *testing.T) {
 	gs.keypair = kr.Bob().(*ed25519.Keypair)
 
 	_, err = gs.validateVoteMessage("", msg)
-	require.ErrorIs(t, err, errors.ErrBlockDoesNotExist)
+	require.ErrorIs(t, err, models.ErrBlockDoesNotExist)
 	authorityID := kr.Alice().Public().(*ed25519.PublicKey).AsBytes()
 	voteMessage := getMessageFromVotesTracker(gs.tracker.votes, fake.Hash(), authorityID)
 	require.Equal(t, msg, voteMessage)
@@ -91,7 +90,7 @@ func TestMessageTracker_SendMessage(t *testing.T) {
 	gs.keypair = kr.Bob().(*ed25519.Keypair)
 
 	_, err = gs.validateVoteMessage("", msg)
-	require.ErrorIs(t, err, errors.ErrBlockDoesNotExist)
+	require.ErrorIs(t, err, models.ErrBlockDoesNotExist)
 	authorityID := kr.Alice().Public().(*ed25519.PublicKey).AsBytes()
 	voteMessage := getMessageFromVotesTracker(gs.tracker.votes, next.Hash(), authorityID)
 	require.Equal(t, msg, voteMessage)
@@ -143,7 +142,7 @@ func TestMessageTracker_ProcessMessage(t *testing.T) {
 	gs.keypair = kr.Bob().(*ed25519.Keypair)
 
 	_, err = gs.validateVoteMessage("", msg)
-	require.ErrorIs(t, err, errors.ErrBlockDoesNotExist)
+	require.ErrorIs(t, err, models.ErrBlockDoesNotExist)
 	authorityID := kr.Alice().Public().(*ed25519.PublicKey).AsBytes()
 	voteMessage := getMessageFromVotesTracker(gs.tracker.votes, next.Hash(), authorityID)
 	require.Equal(t, msg, voteMessage)
