+++ lib/grandpa/models/errors.go
@@ -1,7 +1,7 @@
 // Copyright 2022 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package errors
+package models
 
 import "errors"
