+++ lib/grandpa/vote_message.go
@@ -12,7 +12,6 @@ import (
 	"github.com/ChainSafe/gossamer/dot/telemetry"
 	"github.com/ChainSafe/gossamer/lib/blocktree"
 	"github.com/ChainSafe/gossamer/lib/crypto/ed25519"
-	grandpaerrors "github.com/ChainSafe/gossamer/lib/grandpa/errors"
 	"github.com/ChainSafe/gossamer/lib/grandpa/models"
 	"github.com/ChainSafe/gossamer/pkg/scale"
 
@@ -205,7 +204,7 @@ func (s *Service) validateVoteMessage(from peer.ID, m *models.VoteMessage) (*mod
 	}
 
 	err = s.validateVote(vote)
-	if errors.Is(err, grandpaerrors.ErrBlockDoesNotExist) ||
+	if errors.Is(err, models.ErrBlockDoesNotExist) ||
 		errors.Is(err, blocktree.ErrDescendantNotFound) ||
 		errors.Is(err, blocktree.ErrEndNodeNotFound) ||
 		errors.Is(err, blocktree.ErrStartNodeNotFound) {
@@ -287,7 +286,7 @@ func (s *Service) validateVote(v *models.Vote) error {
 	}
 
 	if !has {
-		return grandpaerrors.ErrBlockDoesNotExist
+		return models.ErrBlockDoesNotExist
 	}
 
 	// check if the block is an eventual descendant of a previously finalised block
